Entrega 6

+++ fuenteProxy/src/main/java/ar/edu/utn/frba/dds/fuenteproxy/services/impl/FuentesService.java
@@ -1,7 +1,9 @@
 package ar.edu.utn.frba.dds.fuenteproxy.services.impl;
 
+import ar.edu.utn.frba.dds.fuenteproxy.domain.dtos.DTOConverter;
 import ar.edu.utn.frba.dds.fuenteproxy.domain.dtos.input.FuenteInputDTO;
 import ar.edu.utn.frba.dds.fuenteproxy.domain.dtos.input.HechoInputDTO;
+import ar.edu.utn.frba.dds.fuenteproxy.domain.dtos.output.FuenteAMostrarOutputDTO;
 import ar.edu.utn.frba.dds.fuenteproxy.domain.dtos.output.FuenteOutputDTO;
 import ar.edu.utn.frba.dds.fuenteproxy.domain.entities.fuentes.Fuente;
 import ar.edu.utn.frba.dds.fuenteproxy.domain.entities.fuentes.FuenteDDS;
@@ -10,13 +12,14 @@ import ar.edu.utn.frba.dds.fuenteproxy.domain.repositories.IFuentesRepositoryJPA
 import ar.edu.utn.frba.dds.fuenteproxy.services.IFuentesService;
 import ar.edu.utn.frba.dds.fuenteproxy.services.IHechosService;
 import lombok.Data;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
+import org.springframework.transaction.annotation.Transactional;
+
 
 import java.util.List;
-import java.util.stream.Collectors;
+
 
 @Data
 @Service
@@ -31,6 +34,12 @@ public class FuentesService implements IFuentesService {
         this.fuentesRepository = fuentesRepository;
         this.fuenteFactory = fuenteFactory;
         this.hechosService = hechosService;
+    }
+
+    @Override
+    public List<FuenteAMostrarOutputDTO> getFuentes(){
+        return fuentesRepository.findAll()
+                .stream().map(DTOConverter::mapToFuenteAMostrarOutputDTO).toList();
     }
 
     @Override
@@ -55,7 +64,7 @@ public class FuentesService implements IFuentesService {
 
     @Override
     public void agregarFuenteDDS(FuenteInputDTO dto) {
-        boolean existsFuenteByBaseUrl = fuentesRepository.existsFuenteByBaseUrl(dto.getBaseUrl());
+        boolean existsFuenteByBaseUrl = fuentesRepository.existsFuenteByBaseUrl(fuenteFactory.getDdsBaseUrl());
         FuenteDDS nuevaFuente = fuenteFactory.nuevaFuenteDDS(dto.getNombre());
         if(!existsFuenteByBaseUrl) {
             hechosService.cargarHechosFuente(nuevaFuente).subscribe();
@@ -65,7 +74,18 @@ public class FuentesService implements IFuentesService {
 
 
     @Override
+    @Transactional
     public void eliminarFuente(String nombre) {
+        Fuente fuente = fuentesRepository.findByNombre(nombre);
+
+        if (fuente == null) {
+            throw new RuntimeException("Fuente no encontrada: " + nombre);
+        }
+
+        if (fuente.getTipo() == TipoFuenteProxy.EXTERNA) {
+            hechosService.eliminarHechosDeFuente(fuente.getId());
+        }
+
         fuentesRepository.deleteByNombre(nombre);
     }
